requirements.txt
Flask
requests
Pillow
pinyin
reportlab
gunicorn
numpy